Almost flawless SideScroller game type is working
Have collision for Up,Down,Left and right sides against blocks for ALL
entities :3 The performance hasn't suffered that much I hope, the engine
is still very going and will be updated. I will be finishing up on the
SideScroller game type and start working a lot more on the RPG side of
things, I think that will be a lot cooler than side scroller. Blocks are
going to be used in both SideScroller game and RPG style games, as will
Custom and "Free-fly" games. Collision and all that will work just as
they do in RPG or SideScroller types.