Successfully added targeting
Now the player is able to target enemies by right clicking, the enemy
will turn red once pressed, blue if it is able to be pressed. If the
enemy is not close to the player anymore after being selected it is
automatically unselected.

Works quite well, although I'll need to add some more to it later on.